+++ crates/flipperzero/src/dialogs/mod.rs
@@ -152,6 +152,12 @@ impl<'a> Drop for DialogMessage<'a> {
     }
 }
 
+impl<'a> Default for DialogMessage<'a> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl DialogMessageButton {
     fn from_sys(sys: sys::DialogMessageButton) -> Option<Self> {
         match sys {