+++ crates/flipperzero/src/lib.rs
@@ -2,8 +2,9 @@
 
 #![no_std]
 #![cfg_attr(test, no_main)]
+#![deny(rustdoc::broken_intra_doc_links)]
 
-#[cfg(feature = "alloc")]
+#[cfg(any(feature = "alloc", docsrs))]
 extern crate alloc;
 
 pub mod dialogs;